# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/productos/categoria/{categoriaId}": {
      "get": {
        "operationId": "ProductosController_findByCategoryPublic",
        "parameters": [
          {
            "name": "categoriaId",
            "required": true,
            "in": "path",
            "description": "ID de la categoría (debe ser un número entero mayor a 0)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener productos por categoría (público)",
        "tags": [
          "Productos"
        ]
      }
    },
...
    "/ocasiones": {
      "post": {
        "operationId": "OcasionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearGenericoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin crea una ocasión",
        "tags": [
          "Ocasiones"
        ]
      },
      "get": {
        "operationId": "OcasionController_findAll",
        "parameters": [
          {
            "name": "nombre",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todo ocasion con filtro opcional por nombre",
        "tags": [
          "Ocasiones"
        ]
      }
    },
    "/ocasiones/{id}": {
      "get": {
        "operationId": "OcasionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener una ocasión por ID",
        "tags": [
          "Ocasiones"
        ]
      },
      "patch": {
        "operationId": "OcasionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarGenericoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar una ocasión por ID",
        "tags": [
          "Ocasiones"
        ]
      },
      "delete": {
        "operationId": "OcasionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar una ocasión por ID",
        "tags": [
          "Ocasiones"
        ]
      }
    },
...
    "/categoria": {
      "get": {
        "operationId": "CategoriaController_getAllPublic",
        "parameters": [
          {
            "name": "nombre",
            "required": false,
            "in": "query",
            "description": "Filtrar las categorías por nombre (opcional)",
            "schema": {
              "example": "anillos",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todas las categorías con filtro opcional por nombre",
        "tags": [
          "Categoria"
        ]
      },
      "post": {
        "operationId": "CategoriaController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearCategoriaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin crea una categoria",
        "tags": [
          "Categoria"
        ]
      }
    },
    "/categoria/{id}": {
      "patch": {
        "operationId": "CategoriaController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarCategoriaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin actualiza una categoria",
        "tags": [
          "Categoria"
        ]
      },
      "delete": {
        "operationId": "CategoriaController_deleteCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin elimina una categoria",
        "tags": [
          "Categoria"
        ]
      }
    },
...
    "/tiendas": {
      "get": {
        "operationId": "TiendaController_getAllPublicas",
        "parameters": [
          {
            "name": "nombre",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Listar todos los estilos con filtro opcional por nombre, se puede consultar asi tambien para todas las caracteristicas /tiendas?nombre=mo",
        "tags": [
          "Tienda"
        ]
      }
    },
    "/tiendas/{id}": {
      "get": {
        "operationId": "TiendaController_getUnaPublica",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Un usuario público puede ver toda la información de una tienda",
        "tags": [
          "Tienda"
        ]
      }
    },
    "/tiendas/mi-tienda": {
      "patch": {
        "operationId": "TiendaController_actualizarTienda",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarTiendaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vendedor actualiza su tienda",
        "tags": [
          "Tienda"
        ]
      }
    },
    "/carrito": {
      "post": {
        "operationId": "CarritoController_agregarProducto",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarCarritoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agregar un producto al carrito o aumentar su cantidad",
        "tags": [
          "Carrito"
        ]
      },
      "get": {
        "operationId": "CarritoController_obtenerCarrito",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todos los productos del carrito del usuario",
        "tags": [
          "Carrito"
        ]
      },
      "delete": {
        "operationId": "CarritoController_vaciarCarrito",
        "parameters": [],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vaciar todo el carrito del usuario",
        "tags": [
          "Carrito"
        ]
      }
    },
    "/carrito/{productoId}/quitar-uno": {
      "patch": {
        "operationId": "CarritoController_quitarUnaUnidad",
        "parameters": [
          {
            "name": "productoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disminuir en una unidad la cantidad de un producto en el carrito",
        "tags": [
          "Carrito"
        ]
      }
    },
    "/carrito/{productoId}": {
      "delete": {
        "operationId": "CarritoController_eliminarProducto",
        "parameters": [
          {
            "name": "productoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar completamente un producto del carrito",
        "tags": [
          "Carrito"
        ]
      }
    },
...
  },
  "info": {
    "title": "MarketJoyas",
    "description": "Documentacion de endpoints de marketjoyas",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CrearGenericoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre de la entidad (unico)",
            "example": "x"
          }
        },
        "required": [
          "nombre"
        ]
      },
      "ActualizarGenericoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre de la entidad (unico)",
            "example": "Moderno"
          }
        }
      },
      "CrearCategoriaDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre de la entidad (único)",
            "example": "x"
          }
        },
        "required": [
          "nombre"
        ]
      },
      "ActualizarCategoriaDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre de la entidad (único)",
            "example": "x"
          }
        },
        "required": [
          "nombre"
        ]
      },
...
      "ActualizarTiendaDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre o razón social de la tienda",
            "example": "Mi Tienda S.A.C."
          },
          "direccion": {
            "type": "string",
            "description": "Dirección fiscal de la tienda",
            "example": "Av. Principal 123, Lima"
          },
          "telefono": {
            "type": "string",
            "description": "Teléfono de contacto de la tienda",
            "example": "+51987654321"
          },
          "pais": {
            "type": "string",
            "description": "País donde opera la tienda",
            "example": "Perú"
          },
          "ciudad": {
            "type": "string",
            "description": "Ciudad donde está ubicada la tienda",
            "example": "Lima"
          },
          "provincia": {
            "type": "string",
            "description": "Provincia de la tienda",
            "example": "Lima"
          },
          "departamento": {
            "type": "string",
            "description": "Departamento de la tienda",
            "example": "Lima"
          },
          "distrito": {
            "type": "string",
            "description": "Distrito de la tienda",
            "example": "San Isidro"
          },
          "codigoPostal": {
            "type": "string",
            "description": "Código postal",
            "example": "15036"
          },
          "ruc": {
            "type": "string",
            "description": "Número de RUC de la tienda",
            "example": "20481234567"
          },
          "emailTienda": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico de la tienda",
            "example": "contacto@mitienda.com"
          },
          "logo": {
            "type": "string",
            "description": "URL del logo de la tienda",
            "example": "https://res.cloudinary.com/tuapp/image/upload/v123456/logo.png"
          },
          "logoPublicId": {
            "type": "string",
            "description": "ID público del logo en Cloudinary",
            "example": "tienda/logo123"
          }
        }
      },
      "AgregarCarritoDto": {
        "type": "object",
        "properties": {
          "productoId": {
            "type": "number",
            "description": "ID del producto a agregar al carrito",
            "example": 5
          },
          "cantidad": {
            "type": "number",
            "minimum": 1,
            "description": "Cantidad del producto (mínimo 1)",
            "example": 2
          }
        },
        "required": [
          "productoId",
          "cantidad"
        ]
      },
...
    }
  }
}